// src/app/master/model/clientmaster.ts
export class ClientMaster {
    dept: string = "";
    employee: string = "";
    compname: string = "";
    clientname: string = "";
    clientorgtype: string = "";
    mobno1: number = 0;
    mobno2: number = 0;
    landline: number = 0;
    email1: string = "";
    email2: string = "";
    gstno: string = "";
    website: string = "";
    status: string = "";
}

export class ClientOrgNameSelect {
    orgname: string = "";
    orgcode: string = "";
}